Refactor Bible tab into smaller composables
Extracted and modularized the Bible tab UI into focused composables and files for clarity and reuse. Added: BibleColumnHeaderRow, BibleColumns, BibleCrossReferences, BibleDetectionPanel, BibleHistoryPanel, BibleTranslationOrder. Updated BibleTab.kt to use these components and removed large duplicated UI blocks (cross-ref popover/panel, detection strip, history list, translation-order UI and header). Adjusted imports, state callbacks and wiring to preserve behaviour. This improves readability, testability and separates presentation concerns without changing user-visible functionality.
